Added in some colombian gear
