+++ @twc_config/addons/twc_colombian/backpacks.hpp
@@ -1,0 +1,3 @@
+#define ADD_MAGA(a,b) class _nc_##a {magazine = a; count = b;}
+#define ADD_WEAP(a,b) class _nc_##a {weapon = a; count = b;}
+#define ADD_ITEM(a,b) class _nc_##a {name = a; count = b;}
