+++ @twc_config/addons/twc_colombian/groups.hpp
@@ -1,0 +1,31 @@
+class TWC_Groups_Colombia
+{
+	name = "Infantry";
+	class Squad
+	{
+		name = "Squad";
+		faction = "TWC_Colombia";
+		side = 1;
+		class Unit0
+		{
+				side = 1;
+				vehicle = "TWC_Infantry_Colombia_Rifleman";
+				rank = "SERGEANT";
+				position[] = {0,0,0};
+		};
+		class Unit1
+		{
+				side = 1;
+				vehicle = "TWC_Infantry_Colombia_Grenadier";
+				rank = "PRIVATE";
+				position[] = {2,0,0};
+		};
+		class Unit2
+		{
+				side = 1;
+				vehicle = "TWC_Infantry_Colombia_MG";
+				rank = "PRIVATE";
+				position[] = {4,0,0};
+		};
+	};
+};
